+++ src/resolver/form/index.ts
@@ -1,5 +1,6 @@
 import { FormCreateMutation } from './form.create.mutation'
 import { FormDeleteMutation } from './form.delete.mutation'
+import { FormFieldResolver } from './form.field.resolver'
 import { FormListQuery } from './form.list.query'
 import { FormQuery } from './form.query'
 import { FormResolver } from './form.resolver'
@@ -11,6 +12,7 @@ import { PageResolver } from './page.resolver'
 export const formResolvers = [
   FormCreateMutation,
   FormDeleteMutation,
+  FormFieldResolver,
   FormQuery,
   FormResolver,
   FormListQuery,