+++ src/service/form/form.create.service.ts
@@ -3,13 +3,16 @@ import { InjectRepository } from '@nestjs/typeorm'
 import { Repository } from 'typeorm'
 import { FormCreateInput } from '../../dto/form/form.create.input'
 import { FormEntity } from '../../entity/form.entity'
+import { PageEntity } from '../../entity/page.entity'
 import { UserEntity } from '../../entity/user.entity'
+import { FormPageCreateService } from './form.page.create.service'
 
 @Injectable()
 export class FormCreateService {
   constructor(
     @InjectRepository(FormEntity)
-    private readonly formRepository: Repository<FormEntity>
+    private readonly formRepository: Repository<FormEntity>,
+    private readonly formPageCreateService: FormPageCreateService,
   ) {
   }
 
@@ -23,9 +26,14 @@ export class FormCreateService {
     form.language = input.language || 'en'
     form.design.layout = input.layout
 
+
+    form.endPage = this.formPageCreateService.create(input.endPage)
+    form.startPage = this.formPageCreateService.create(input.startPage)
+
     form.admin = admin
 
     return await this.formRepository.save(form)
   }
+
 }